pyproject.toml
# Project metadata
[project]
name = "blog-with-FastApi-and-MongoDB-ODM"
version = "1.0.0"
description = ""
authors = [{ name = "Nayan Biswas", email = "nayan32biswas@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"

# Main dependencies as an array
dependencies = [
    "bcrypt>=4.1.2",
    "cryptography>=44.0.2",
    "fastapi>=0.109.0",
    "gunicorn>=23.0.0",
    "mongodb-odm>=1.0.0",
    "passlib>=1.7.4",
    "pydantic>=2.5.3",
    "pyjwt>=2.8.0",
    "pymongo[srv]>=4.6.1",
    "python-multipart>=0.0.6",
    "python-slugify>=8.0.2",
    "typer>=0.9.0",
    "types-python-slugify>=8.0.2.20240127",
    "uvicorn>=0.27.0.post1",
]

# Optional dev dependencies
[project.optional-dependencies]
dev = [
    "coverage[toml]>=7.4.1",
    "debugpy>=1.8.14",
    "Faker>=22.5.1",
    "httpx>=0.26.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "pytest>=8.0.0",
    "ruff>=0.1.14",
    "types-passlib>=1.7.7.20240106",
    "types-python-slugify>=8.0.2.20240310",
]

# Optional uv-specific configuration
[tool.uv]

[tool.uv.sources]
mongodb-odm = { git = "https://github.com/nayan32biswas/mongodb-odm", rev = "main" }
# constraint = "constraints.txt"

# Tool configurations
[tool.coverage.run]
parallel = true
context = '${CONTEXT}'
omit = ["app/tests/**", "cli/**"]

[tool.coverage.report]
exclude_lines = ['if __name__ == "__main__":']

[tool.mypy]
strict = true

[[tool.mypy.overrides]]
module = ""
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = ""
warn_unused_ignores = false

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C4", "B", "UP"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-third-party = ["mongodb_odm", "pydantic", "fastapi"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
